Move `box_plot` to core for re-use
Did not yet move the one from `performance` because it is slightly
different and I am half asleep - i'll tackle when i'm fresh
